Added xFuturePosition and yFuturePosition properties to the Projectile class which are to be used for better collision detection.

+++ TronDuel/TronDuel/GraphicalObjects/MovingObjects/SpaceShip.cs
@@ -93,13 +93,13 @@
 
         private void EraseSpriteFromLastPosition()
         {
-            Console.SetCursorPosition((int)this.Xposition, (int)this.Yposition);
+            Console.SetCursorPosition((byte)this.Xposition, (byte)this.Yposition);
             Console.Write(' ');
         }
 
         public override void Draw()
         {
-            Console.SetCursorPosition((int)this.Xposition, (int)this.Yposition);
+            Console.SetCursorPosition((byte)this.Xposition, (byte)this.Yposition);
             Console.ForegroundColor = this.Color;
             Console.Write(this.Sprite);
         }
